# 默认模块.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "默认模块",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/certification/apply": {
      "post": {
        "summary": "提交认证申请",
        "deprecated": false,
        "description": "提交认证申请",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCertificationRequest",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultLong"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": 0
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/certification/status/{applyType}": {
      "get": {
        "summary": "获取用户认证申请状态",
        "deprecated": false,
        "description": "获取用户认证申请状态",
        "tags": [],
        "parameters": [
          {
            "name": "applyType",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultUserCertificationApply"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "id": 0,
                    "userId": 0,
                    "applyType": "",
                    "idNumber": "",
                    "idCardFrontUrl": "",
                    "idCardBackUrl": "",
                    "businessLicenseUrl": "",
                    "legalRepresentative": "",
                    "applyReason": "",
                    "status": "",
                    "adminRemark": "",
                    "approvedTime": "",
                    "createTime": "",
                    "updateTime": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/reviews/page": {
      "get": {
        "summary": "分页查询评价",
        "deprecated": false,
        "description": "分页查询评价",
        "tags": [],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "targetId",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageCooperationReview"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderId": 0,
                        "reviewFrom": "",
                        "reviewTo": "",
                        "targetId": 0,
                        "rating": 0,
                        "comment": "",
                        "createTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/reviews/my": {
      "get": {
        "summary": "查询我的评价",
        "deprecated": false,
        "description": "查询我的评价",
        "tags": [],
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageCooperationReview"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderId": 0,
                        "reviewFrom": "",
                        "reviewTo": "",
                        "targetId": 0,
                        "rating": 0,
                        "comment": "",
                        "createTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/reviews/received": {
      "get": {
        "summary": "读替的评价",
        "deprecated": false,
        "description": "读曏的评价",
        "tags": [],
        "parameters": [
          {
            "name": "targetId",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "current",
            "in": "query",
            "description": "",
            "required": true,
            "example": 1,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "",
            "required": true,
            "example": 10,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultPageCooperationReview"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": {
                    "records": [
                      {
                        "id": 0,
                        "orderId": 0,
                        "reviewFrom": "",
                        "reviewTo": "",
                        "targetId": 0,
                        "rating": 0,
                        "comment": "",
                        "createTime": ""
                      }
                    ],
                    "total": 0,
                    "size": 0,
                    "current": 0,
                    "orders": [
                      {
                        "column": "",
                        "asc": false
                      }
                    ],
                    "optimizeCountSql": false,
                    "searchCount": false,
                    "optimizeJoinOfCountSql": false,
                    "maxLimit": 0,
                    "countId": ""
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/reviews/{reviewId}": {
      "put": {
        "summary": "修改评价",
        "deprecated": false,
        "description": "修改评价",
        "tags": [],
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "rating",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "comment",
            "in": "query",
            "description": "",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除评价",
        "deprecated": false,
        "description": "删除评价",
        "tags": [],
        "parameters": [
          {
            "name": "reviewId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultVoid"
                },
                "example": {
                  "code": 0,
                  "message": "",
                  "data": null
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultLong": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "type": "integer",
            "description": "",
            "format": "int64"
          }
        }
      },
...
      "UserCertificationRequest": {
        "type": "object",
        "properties": {
          "applyType": {
            "type": "string",
            "description": "申请类型：farmer(农户) purchaser(采购方)"
          },
          "idNumber": {
            "type": "string",
            "description": "身份证号或营业执照号"
          },
          "idCardFrontUrl": {
            "type": "string",
            "description": "身份证正面照URL"
          },
          "idCardBackUrl": {
            "type": "string",
            "description": "身份证反面照URL"
          },
          "businessLicenseUrl": {
            "type": "string",
            "description": "营业执照照片URL"
          },
          "legalRepresentative": {
            "type": "string",
            "description": "法定代表人"
          },
          "applyReason": {
            "type": "string",
            "description": "申请说明"
          }
        },
        "required": [
          "applyType",
          "idNumber",
          "idCardFrontUrl"
        ]
      },
...
      "UserCertificationApply": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "userId": {
            "type": "integer",
            "description": "申请用户ID",
            "format": "int64"
          },
          "applyType": {
            "type": "string",
            "description": "申请类型：farmer(农户) purchaser(采购方)"
          },
          "idNumber": {
            "type": "string",
            "description": "身份证号或营业执照号"
          },
          "idCardFrontUrl": {
            "type": "string",
            "description": "身份证正面照URL"
          },
          "idCardBackUrl": {
            "type": "string",
            "description": "身份证反面照URL"
          },
          "businessLicenseUrl": {
            "type": "string",
            "description": "营业执照照片URL"
          },
          "legalRepresentative": {
            "type": "string",
            "description": "法定代表人"
          },
          "applyReason": {
            "type": "string",
            "description": "申请说明"
          },
          "status": {
            "type": "string",
            "description": "申请状态：pending(待审核) approved(已批准) rejected(已拒绝)"
          },
          "adminRemark": {
            "type": "string",
            "description": "管理员备注"
          },
          "approvedTime": {
            "type": "string",
            "description": "批准时间"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间"
          }
        }
      },
      "ResultVoid": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "description": "",
            "type": "null"
          }
        }
      },
      "ResultUserCertificationApply": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/UserCertificationApply",
            "description": ""
          }
        }
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "column": {
            "type": "string",
            "description": "需要进行排序的字段"
          },
          "asc": {
            "type": "boolean",
            "description": "是否正序排列，默认 true",
            "default": true
          }
        }
      },
...
      "CooperationReview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "reviewFrom": {
            "type": "string",
            "description": ""
          },
          "reviewTo": {
            "type": "string",
            "description": ""
          },
          "targetId": {
            "type": "integer",
            "description": "",
            "format": "int64"
          },
          "rating": {
            "type": "integer",
            "description": ""
          },
          "comment": {
            "type": "string",
            "description": ""
          },
          "createTime": {
            "type": "string",
            "description": ""
          }
        }
      },
      "PageCooperationReview": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CooperationReview",
              "description": "合作评价实体类"
            },
            "description": "查询数据列表",
            "default": "Collections.emptyList()"
          },
          "total": {
            "type": "integer",
            "description": "总数",
            "format": "int64",
            "default": 0
          },
          "size": {
            "type": "integer",
            "description": "每页显示条数，默认 10",
            "format": "int64",
            "default": 10
          },
          "current": {
            "type": "integer",
            "description": "当前页",
            "format": "int64",
            "default": 1
          },
          "orders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem",
              "description": "com.baomidou.mybatisplus.core.metadata.OrderItem"
            },
            "description": "排序字段信息",
            "default": "new ArrayList<>()"
          },
          "optimizeCountSql": {
            "type": "boolean",
            "description": "自动优化 COUNT SQL",
            "default": true
          },
          "searchCount": {
            "type": "boolean",
            "description": "是否进行 count 查询",
            "default": true
          },
          "optimizeJoinOfCountSql": {
            "type": "boolean",
            "description": "{@link #optimizeJoinOfCountSql()}",
            "default": true
          },
          "maxLimit": {
            "type": "integer",
            "description": "单页分页条数限制",
            "format": "int64"
          },
          "countId": {
            "type": "string",
            "description": "countId"
          }
        }
      },
      "ResultPageCooperationReview": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": ""
          },
          "message": {
            "type": "string",
            "description": ""
          },
          "data": {
            "$ref": "#/components/schemas/PageCooperationReview",
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}